python/bert-extractive-summarizer/requirements.txt
spacy==2.1.3
bert-extractive-summarizer>=0.1.4
beautifulsoup4
